perf: append shared string indexes without temporaries

+++ include/fastxlsx/detail/xml.hpp
@@ -17,7 +17,7 @@ void append_escaped_xml_text(std::string& output, std::string_view value);
 void append_escaped_xml_attribute(std::string& output, std::string_view value);
 [[nodiscard]] std::string format_number(double value);
 void append_number(std::string& output, double value);
-void append_unsigned_decimal(std::string& output, std::uint32_t value);
+void append_unsigned_decimal(std::string& output, std::uint64_t value);
 void append_cell_reference(std::string& output, std::uint32_t row, std::uint32_t column);
 [[nodiscard]] std::string cell_reference(std::uint32_t row, std::uint32_t column);
 [[nodiscard]] std::string range_reference(std::uint32_t first_row,
